1919 - Refactor form
